make rc filter processing width configurable
because rounding errors in the RC filter accumulated too big, there was
the decision to seperate interface width and processing width.

a rule of thumb is: be 20dB better than your incoming signal, therefore
win in dB = N*6dB + 3dB
and we should be at least 3 bit better in processing.

for double safety, we use 4 bits in the testbench and the data slicer
from now on.

testbench final value error went from >20LSB to 1LSB